lazer base behavior (bugged)

+++ Assets/_Scripts/Enemy/EnemyHitbox.cs
@@ -9,7 +9,7 @@ public class EnemyHitbox : MonoBehaviour
     {
         _singleEnemyManager = transform.GetComponent<SingleEnemyManager>();
     }
-    public void OnHit(int baseDamage)
+    public void OnHit(float baseDamage)
     {
         //Debug.Log(transform.name+" got hit");
         _singleEnemyManager.OnHit(baseDamage);
